+++ NEWS.md
@@ -1,5 +1,8 @@
 # nanoscopeAFM 2.0.12
 
+* read images that are not square
+* update NanoScope image reader for V8
+* `AFM.dataType()` has new type of "spectroscopy"
 * Use `raster::click()` function to select two points in `AFM.lineProfile()` if coordinates are not provided
 * Support pixels for profile scans in `AFM.lineProfile()`